feature: cleaned up user ingester
